move lesson to module
